// source_code/types/achievement.ts
export interface AchievementFormData {
  title: string;
  category: string;
  difficulty: string;
  description: string;
  image: string;
  points: string;
  expiration: string;
  metadata: string;
}

export const achievementFormDefault: AchievementFormData = {
  title: "",
  category: "Education",
  difficulty: "Beginner",
  description: "",
  image: "",
  points: "",
  expiration: "",
  metadata: "",
};

export interface Achievement {
  id: string;

  issuer_id: string;

  title: string;

  category: string | null;

  difficulty: string | null;

  description: string | null;

  image: string | null;

  points: number | null;

  expiration: string | null;

  metadata: {
    note: string;
  } | null;

  created_at: string;

  updated_at: string;

  status: "published" | "archived";

  archived_at: string | null;
}